backend/requirements.txt
fastapi
openai
python-dotenv
uvicorn 